requirements.txt
flask
line-bot-sdk
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
